move metadata protobuffers into proto/
We are attempting to have a single protobuffer package/namespace and
therefore need to move the metadata-specific protobuffer messages (back)
into the $ROOT/proto directory (and the .proto definition files into the
$ROOT/proto/defs directory).

This patch simply does that, mostly just changing package import paths.

Issue #111
