+++ Makefile
@@ -3,8 +3,8 @@ VENDOR := vendor
 VERSION := $(shell git describe --tags --always --dirty)
 API_PROTO_DIR := $(shell pwd)/pkg/api/proto
 API_PROTO_DEFS_DIR := $(shell pwd)/pkg/api/proto/defs
-META_PROTO_DIR := $(shell pwd)/pkg/metadata/proto
-META_PROTO_DEFS_DIR := $(shell pwd)/pkg/metadata/proto/defs
+META_PROTO_DIR := $(shell pwd)/proto
+META_PROTO_DEFS_DIR := $(shell pwd)/proto/defs
 RESOURCE_PROTO_DIR := $(shell pwd)/pkg/resource/proto
 RESOURCE_PROTO_DEFS_DIR := $(shell pwd)/pkg/resource/proto/defs
 GO_BIN_DIR := $(GOPATH)/bin
