+++ pkg/api/server/server.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/runmachine-io/runmachine/pkg/api/server/config"
 	"github.com/runmachine-io/runmachine/pkg/logging"
-	metapb "github.com/runmachine-io/runmachine/pkg/metadata/proto"
 	respb "github.com/runmachine-io/runmachine/pkg/resource/proto"
+	metapb "github.com/runmachine-io/runmachine/proto"
 )
 
 var (
