+++ pkg/metadata/conditions/object.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strconv"
 
-	pb "github.com/runmachine-io/runmachine/pkg/metadata/proto"
+	pb "github.com/runmachine-io/runmachine/proto"
 )
 
 // ObjectCondition is a class used in filtering objects.  Optional partition
