+++ pkg/metadata/conditions/partition.go
@@ -1,6 +1,6 @@
 package conditions
 
-import pb "github.com/runmachine-io/runmachine/pkg/metadata/proto"
+import pb "github.com/runmachine-io/runmachine/proto"
 
 type HasPartition interface {
 	GetPartition() string
