+++ pkg/metadata/conditions/property.go
@@ -1,7 +1,7 @@
 package conditions
 
 import (
-	pb "github.com/runmachine-io/runmachine/pkg/metadata/proto"
+	pb "github.com/runmachine-io/runmachine/proto"
 )
 
 type HasProperties interface {
