+++ pkg/metadata/types/object.go
@@ -1,7 +1,7 @@
 package types
 
 import (
-	pb "github.com/runmachine-io/runmachine/pkg/metadata/proto"
+	pb "github.com/runmachine-io/runmachine/proto"
 )
 
 type ObjectMatcher interface {
