+++ proto/defs/filter.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package runmmeta;
+package runm;
 
 message UuidFilter {
     string uuid = 1;
