+++ proto/defs/object_definition.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package runmmeta;
+package runm;
 
 import "property.proto";
