+++ proto/defs/provider_type.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package runmmeta;
+package runm;
 
 import "filter.proto";
