+++ proto/defs/search.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package runmmeta;
+package runm;
 
 enum SortDirection {
     ASC = 0;
