--- proto/defs/service_account.proto
@@ -1,274 +1,0 @@
-syntax = "proto3";
-
-package runm;
-
-import "error.proto";
-import "permission.proto";
-import "project.proto";
-import "quota.proto";
-import "role.proto";
-import "search.proto";
-import "session.proto";
-import "user.proto";
-import "wrappers.proto";
-
-// The runm-account gRPC service handles operations for users,
-// projects, quota limits, etc.
-service RunmAccount {
-    // Returns information about a specific role
-    rpc role_get(RoleGetRequest) returns (Role) {}
-
-    // Deletes one or more roles
-    rpc role_delete(RoleDeleteRequest) returns (RoleDeleteResponse) {}
-
-    // Set information about a specific role
-    rpc role_set(RoleSetRequest) returns (RoleSetResponse) {}
-
-    // Returns information about multiple roles
-    rpc role_list(RoleListRequest) returns (stream Role) {}
-
-    // Returns information about a specific user
-    rpc user_get(UserGetRequest) returns (User) {}
-
-    // Deletes one or more users
-    rpc user_delete(UserDeleteRequest) returns (UserDeleteResponse) {}
-
-    // Set information about a specific user
-    rpc user_set(UserSetRequest) returns (UserSetResponse) {}
-
-    // Returns information about multiple users
-    rpc user_list(UserListRequest) returns (stream User) {}
-
-    // List projects a user belongs to
-    rpc user_members_list(UserMembersListRequest) returns (stream Project) {}
-
-    // List roles a user has
-    rpc user_roles_list(UserRolesListRequest) returns (stream Role) {}
-
-    // Add or remove roles from a user
-    rpc user_roles_set(UserRolesSetRequest) returns (UserRolesSetResponse) {}
-
-    // Returns information about a specific project
-    rpc project_get(ProjectGetRequest) returns (Project) {}
-
-    // Deletes one or more projects
-    rpc project_delete(ProjectDeleteRequest) returns (ProjectDeleteResponse) {}
-
-    // Set information about a specific project
-    rpc project_set(ProjectSetRequest) returns (ProjectSetResponse) {}
-
-    // Returns information about multiple projects
-    rpc project_list(ProjectListRequest) returns (stream Project) {}
-
-    // Add or remove users from a project
-    rpc project_members_set(ProjectMembersSetRequest) returns (ProjectMembersSetResponse) {}
-
-    // List users belonging to a project
-    rpc project_members_list(ProjectMembersListRequest) returns (stream User) {}
-
-    // List quota limits for n project
-    rpc project_quotas_list(ProjectQuotasListRequest) returns (stream ProjectQuota) {}
-
-    // Add, remove or modify quotas from a project
-    rpc project_quotas_set(ProjectQuotasSetRequest) returns (ProjectQuotasSetResponse) {}
-}
-
-// RPC Request payload messages
-message RoleGetRequest {
-    Session session = 1;
-    string search = 2;
-}
-
-message RoleSetFields {
-    StringValue display_name = 1;
-    StringValue project = 2;
-    repeated Permission add = 3;
-    repeated Permission remove = 4;
-}
-
-message RoleSetRequest {
-    Session session = 1;
-    Role role = 2;
-    RoleSetFields changed = 3;
-}
-
-message RoleSetResponse {
-    repeated Error errors = 1;
-    Role role = 2;
-}
-
-message RoleListFilters {
-    repeated string identifiers = 1;
-}
-
-message RoleListRequest {
-    Session session = 1;
-    RoleListFilters filters = 2;
-    SearchOptions options = 3;
-}
-
-message RoleDeleteRequest {
-    Session session = 1;
-    repeated Role roles = 2;
-}
-
-message RoleDeleteResponse {
-    repeated Error errors = 1;
-    uint64 num_deleted = 2;
-}
-
-message UserGetRequestOptions {
-    bool with_roles = 1;
-}
-
-message UserGetRequest {
-    Session session = 1;
-    string search = 2;
-    UserGetRequestOptions options = 100;
-}
-
-message UserSetFields {
-    StringValue email = 1;
-    StringValue display_name = 2;
-    repeated Role roles = 101;
-}
-
-message UserSetRequest {
-    Session session = 1;
-    StringValue search = 2;
-    UserSetFields changed = 3;
-}
-
-message UserSetResponse {
-    repeated Error errors = 1;
-    User user = 2;
-}
-
-message UserListFilters {
-    repeated string identifiers = 1;
-}
-
-message UserListRequest {
-    Session session = 1;
-    UserListFilters filters = 2;
-    SearchOptions options = 3;
-}
-
-message UserMembersListRequest {
-    Session session = 1;
-    User user = 2;
-}
-
-message UserRolesListRequest {
-    Session session = 1;
-    User user = 2;
-}
-
-message UserRolesSetRequest {
-    Session session = 1;
-    User user = 2;
-    repeated string add = 3;
-    repeated string remove = 4;
-}
-
-message UserRolesSetResponse {
-    repeated Error errors = 1;
-    uint64 num_added = 2;
-    uint64 num_removed = 3;
-}
-
-message UserDeleteRequest {
-    Session session = 1;
-    repeated User users = 2;
-}
-
-message UserDeleteResponse {
-    repeated Error errors = 1;
-    uint64 num_deleted = 2;
-}
-
-message ProjectGetRequest {
-    Session session = 1;
-    string search = 2;
-}
-
-message ProjectSetFields {
-    StringValue display_name = 1;
-    StringValue parent = 2;
-}
-
-message ProjectSetRequest {
-    Session session = 1;
-    StringValue search = 2;
-    ProjectSetFields changed = 3;
-}
-
-message ProjectSetResponse {
-    repeated Error errors = 1;
-    Project project = 2;
-}
-
-message ProjectListFilters {
-    repeated string identifiers = 1;
-}
-
-message ProjectListRequest {
-    Session session = 1;
-    ProjectListFilters filters = 2;
-    SearchOptions options = 3;
-}
-
-message ProjectMembersSetRequest {
-    Session session = 1;
-    Project project = 2;
-    repeated string add = 3;
-    repeated string remove = 4;
-}
-
-message ProjectMembersSetResponse {
-    repeated Error errors = 1;
-    uint64 num_added = 2;
-    uint64 num_removed = 3;
-}
-
-message ProjectMembersListRequest {
-    Session session = 1;
-    Project project = 2;
-}
-
-message ProjectDeleteRequest {
-    Session session = 1;
-    repeated Project projects = 2;
-}
-
-message ProjectDeleteResponse {
-    repeated Error errors = 1;
-    uint64 num_deleted = 2;
-}
-
-message ProjectQuotasListFilters {
-    repeated string projects = 1;
-    repeated string resource_types = 3;
-}
-
-message ProjectQuotasListRequest {
-    Session session = 1;
-    ProjectQuotasListFilters filters = 2;
-    SearchOptions options = 3;
-}
-
-message ProjectQuotasSetRequest {
-    Session session = 1;
-    Project project = 2;
-    repeated ProjectQuota add = 3;
-    repeated ProjectQuota remove = 4;
-    repeated ProjectQuota modify = 5;
-}
-
-message ProjectQuotasSetResponse {
-    repeated Error errors = 1;
-    Project project = 2;
-    uint64 num_added = 3;
-    uint64 num_removed = 4;
-    uint64 num_modified = 5;
-}
