+++ proto/defs/service_metadata.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package runmmeta;
+package runm;
 
 import "object.proto";
 import "object_definition.proto";
